update to 3.8-dev.
The change in alias renaming is suspicious; it might be a bug in TS 3.8
(at least a spec bug).

I'm pretty sure that 3.8 should expose some new API things for
convenience at least, but I don't know what those are yet. I'll have to
look at the list of breaking changes or diff the API baselines.